+++ src/drivers/base-cache-driver.ts
@@ -738,8 +738,6 @@ export abstract class BaseCacheDriver<
     try {
       return structuredClone(value);
     } catch (error) {
-      console.log(value);
-
       this.logError(
         `Failed to clone cached value for ${key}, typeof value: ${typeof value}`,
         error,